Moving error handling into its own folder
