+++ errorHandling/IExceptionAsHttpResponse.ts
