+++ index.ts
@@ -1,9 +1,9 @@
 export * from './logger/LoggerService';
 export * from './logger/LoggerServiceV2';
 export * from './logger/HttpException';
-export * from './logger/CommonException';
+export * from './errorHandling/CommonException';
 export * from './logger/ErrorFactory';
-export * from './logger/ErrorFactoryV2';
+export * from './errorHandling/ErrorFactoryV2';
 export * from './nest/GlobalErrorFilter';
 export * from './nest/GlobalErrorFilterV2';
 export * from './apm/ApmHelper';
